% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.hm.R
\docType{package}
\name{PAMhm}
\alias{PAMhm}
\title{Generate Heatmaps Based on Partioning Around Medoids (PAM)}
\description{
Data are partitioned (clustered) into _k_ clusters _around medoids_, a more robust version of K-means and plotted as a fractioned heatmap.
}
\details{
\tabular{ll}{
Package: \tab PAMhm\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2011-01-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>, Henri Sara <henri.sara@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
