% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.hm.R
\docType{package}
\name{PAMhm}
\alias{PAMhm}
\title{Generate Heatmaps Based on Partioning Around Medoids (PAM)}
\description{
Data are partitioned (clustered) into k clusters "around medoids", which is
a more robust version of K-means implemented in the function pam() in the 'cluster' package.
The PAM algorithm is described in Kaufman and Rousseeuw (1990) <doi:10.1002/9780470316801>.
Please refer to the pam() function documentation for more references.
Clustered data is plotted as a split heatmap allowing visualisation of representative
"group-clusters" (medoids) in the data as separated fractions of the graph while those
"sub-clusters" are visualised as a traditional heatmap based on hierarchical clustering.
}
\details{
\tabular{ll}{
Package: \tab PAMhm\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2011-01-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>, Henri Sara <henri.sara@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{package}
